+++ src/edu/ucsc/refactor/internal/detectors/UnusedImports.java
@@ -4,7 +4,7 @@ import edu.ucsc.refactor.Context;
 import edu.ucsc.refactor.internal.visitors.ImportsReferencesVisitor;
 import edu.ucsc.refactor.spi.IssueDetector;
 import edu.ucsc.refactor.spi.Smell;
-import edu.ucsc.refactor.util.AstUtil;
+import edu.ucsc.refactor.internal.util.AstUtil;
 import org.eclipse.jdt.core.dom.ASTNode;
 import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.eclipse.jdt.core.dom.ImportDeclaration;
