+++ src/edu/ucsc/refactor/internal/detectors/UnusedTypes.java
@@ -4,7 +4,7 @@ import edu.ucsc.refactor.Context;
 import edu.ucsc.refactor.internal.visitors.TypeDeclarationVisitor;
 import edu.ucsc.refactor.spi.IssueDetector;
 import edu.ucsc.refactor.spi.Smell;
-import edu.ucsc.refactor.util.AstUtil;
+import edu.ucsc.refactor.internal.util.AstUtil;
 import org.eclipse.jdt.core.dom.*;
 
 import java.util.List;
