+++ src/edu/ucsc/refactor/internal/visitors/LabelVisitor.java
@@ -1,6 +1,6 @@
 package edu.ucsc.refactor.internal.visitors;
 
-import edu.ucsc.refactor.util.AstUtil;
+import edu.ucsc.refactor.internal.util.AstUtil;
 import org.eclipse.jdt.core.dom.*;
 
 import java.util.ArrayList;
