+++ test/edu/ucsc/refactor/internal/InternalTestSuite.java
@@ -2,6 +2,7 @@ package edu.ucsc.refactor.internal;
 
 import edu.ucsc.refactor.internal.changers.ChangersTest;
 import edu.ucsc.refactor.internal.detectors.DetectorsTest;
+import edu.ucsc.refactor.internal.util.ASTUtilTest;
 import junit.framework.Test;
 import junit.framework.TestSuite;
 import org.junit.runner.RunWith;
@@ -13,7 +14,8 @@ import org.junit.runners.Suite;
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
         DetectorsTest.class,
-        ChangersTest.class
+        ChangersTest.class,
+        ASTUtilTest.class
 })
 public class InternalTestSuite {
     public static Test suite() {
