+++ test/edu/ucsc/refactor/util/UtilSuiteTest.java
@@ -12,7 +12,6 @@ import org.junit.runners.Suite;
 @Suite.SuiteClasses({
         StringsTest.class,
         LocationsTest.class,
-        ASTUtilTest.class,
         IoTest.class,
         NotesTest.class,
         CommitHistoryTest.class
